Add macro snapshot tests
Most of our macros don't have explicit tests. This patch adds
a brittle alternative that attempts to render each untested macro
in a special hacky environment and compares its output to a
previously recorded "snapshot". This is very imperfect, but it
may catch accidental changes to macro .ejs files or to the macro
rendering code that would have otherwise gone unnoticed.

The patch also includes a bugfix for the CSS_Ref macro